+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Support for multiple access token signing keys: https://github.com/supertokens/supertokens-core/issues/305
 -   Supporting CDI 2.9
+-   Header `"content-type": "application/json; charset=utf-8"` when querying the core
 
 ### Removed